esercitazionenephila/requirements.txt
asarPy==1.0.1
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
